Makes realtion between channel and group optional.
Also removes is_common field, and the goal is to enable channels
with no belonging group which are common, but not obligatory to
join.

Also starts making a view where one can browse and eventually join
all common channels in which one is not (yet) a member.

This will give larger flexibility of the forum, enabling channels
with different themes, e.g. subjects, events etc.